Refactor. Move package common to domain package.
