+++ src/main/java/com/unvise/bankingsystemapp/domain/account/account/AccountService.java
@@ -1,7 +1,7 @@
 package com.unvise.bankingsystemapp.domain.account.account;
 
 import com.unvise.bankingsystemapp.domain.account.web.dto.AccountDto;
-import com.unvise.bankingsystemapp.common.BaseService;
+import com.unvise.bankingsystemapp.domain.common.BaseService;
 import com.unvise.bankingsystemapp.domain.currency.enums.CurrencyType;
 
 import java.util.List;
