+++ src/main/java/com/unvise/bankingsystemapp/domain/account/history/AccountHistoryService.java
@@ -1,7 +1,7 @@
 package com.unvise.bankingsystemapp.domain.account.history;
 
 import com.unvise.bankingsystemapp.domain.account.web.dto.AccountHistoryDto;
-import com.unvise.bankingsystemapp.common.BaseService;
+import com.unvise.bankingsystemapp.domain.common.BaseService;
 
 public interface AccountHistoryService extends BaseService<AccountHistoryDto, Long> {
 }
