+++ src/main/java/com/unvise/bankingsystemapp/domain/account/web/dto/AccountDto.java
@@ -2,7 +2,7 @@ package com.unvise.bankingsystemapp.domain.account.web.dto;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.unvise.bankingsystemapp.common.View;
+import com.unvise.bankingsystemapp.domain.common.View;
 import com.unvise.bankingsystemapp.domain.currency.enums.CurrencyType;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
