+++ src/main/java/com/unvise/bankingsystemapp/domain/account/web/dto/AccountSecurityDetailsDto.java
@@ -1,7 +1,7 @@
 package com.unvise.bankingsystemapp.domain.account.web.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.unvise.bankingsystemapp.common.View;
+import com.unvise.bankingsystemapp.domain.common.View;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
