+++ src/main/java/com/unvise/bankingsystemapp/domain/auth/web/dto/SignUpDto.java
@@ -1,6 +1,6 @@
 package com.unvise.bankingsystemapp.domain.auth.web.dto;
 
-import com.unvise.bankingsystemapp.common.View;
+import com.unvise.bankingsystemapp.domain.common.View;
 import com.unvise.bankingsystemapp.domain.auth.web.validator.MatchedField;
 import com.unvise.bankingsystemapp.domain.currency.enums.CurrencyType;
 import lombok.AllArgsConstructor;
