+++ src/main/java/com/unvise/bankingsystemapp/domain/common/HomeController.java
@@ -1,4 +1,4 @@
-package com.unvise.bankingsystemapp.common;
+package com.unvise.bankingsystemapp.domain.common;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
