+++ src/main/java/com/unvise/bankingsystemapp/domain/credit/CreditRepository.java
@@ -1,6 +1,6 @@
 package com.unvise.bankingsystemapp.domain.credit;
 
-import com.unvise.bankingsystemapp.common.BaseRepository;
+import com.unvise.bankingsystemapp.domain.common.BaseRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
