+++ src/main/java/com/unvise/bankingsystemapp/domain/credit/CreditService.java
@@ -1,6 +1,6 @@
 package com.unvise.bankingsystemapp.domain.credit;
 
-import com.unvise.bankingsystemapp.common.BaseService;
+import com.unvise.bankingsystemapp.domain.common.BaseService;
 import com.unvise.bankingsystemapp.domain.credit.web.dto.CreditDto;
 import org.springframework.stereotype.Service;
