+++ src/main/java/com/unvise/bankingsystemapp/domain/currency/ExchangeRateService.java
@@ -1,6 +1,6 @@
 package com.unvise.bankingsystemapp.domain.currency;
 
-import com.unvise.bankingsystemapp.common.BaseService;
+import com.unvise.bankingsystemapp.domain.common.BaseService;
 import com.unvise.bankingsystemapp.domain.currency.enums.CurrencyType;
 import com.unvise.bankingsystemapp.domain.currency.web.dto.ExchangeRateDto;
 import com.unvise.bankingsystemapp.exception.resource.ResourceNotFoundException;
