+++ src/main/java/com/unvise/bankingsystemapp/domain/currency/web/dto/ExchangeRateDto.java
@@ -1,7 +1,7 @@
 package com.unvise.bankingsystemapp.domain.currency.web.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.unvise.bankingsystemapp.common.View;
+import com.unvise.bankingsystemapp.domain.common.View;
 import com.unvise.bankingsystemapp.domain.currency.enums.CurrencyType;
 import com.unvise.bankingsystemapp.domain.currency.web.validator.FieldUnmatched;
 import lombok.AllArgsConstructor;
