+++ src/main/java/com/unvise/bankingsystemapp/domain/deposit/DepositService.java
@@ -1,6 +1,6 @@
 package com.unvise.bankingsystemapp.domain.deposit;
 
-import com.unvise.bankingsystemapp.common.BaseService;
+import com.unvise.bankingsystemapp.domain.common.BaseService;
 import com.unvise.bankingsystemapp.domain.deposit.web.dto.DepositDto;
 
 public interface DepositService extends BaseService<DepositDto, Long> {
