+++ src/main/java/com/unvise/bankingsystemapp/domain/person/role/RoleRepository.java
@@ -1,6 +1,6 @@
 package com.unvise.bankingsystemapp.domain.person.role;
 
-import com.unvise.bankingsystemapp.common.BaseRepository;
+import com.unvise.bankingsystemapp.domain.common.BaseRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
