+++ src/main/java/com/unvise/bankingsystemapp/domain/transaction/transaction/TransactionService.java
@@ -1,6 +1,6 @@
 package com.unvise.bankingsystemapp.domain.transaction.transaction;
 
-import com.unvise.bankingsystemapp.common.BaseService;
+import com.unvise.bankingsystemapp.domain.common.BaseService;
 import com.unvise.bankingsystemapp.domain.transaction.web.dto.TransactionDto;
 
 public interface TransactionService extends BaseService<TransactionDto, Long> {
