+++ src/main/java/com/unvise/bankingsystemapp/domain/transaction/web/dto/TransactionDto.java
@@ -1,7 +1,7 @@
 package com.unvise.bankingsystemapp.domain.transaction.web.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.unvise.bankingsystemapp.common.View;
+import com.unvise.bankingsystemapp.domain.common.View;
 import com.unvise.bankingsystemapp.domain.transaction.web.validator.ValidateTransaction;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
